+++ src/Shoply.Domain/DTO/Module/Registration/Product/ProductImage/ExternalPropertiesProductImageDTO.cs
@@ -5,12 +5,12 @@ namespace Shoply.Domain.DTO.Module.Registration;
 public class ExternalPropertiesProductImageDTO : BaseExternalPropertiesDTO<ExternalPropertiesProductImageDTO>
 {
     public string FileName { get; private set; }
-    public decimal FileLength { get; private set; }
+    public long FileLength { get; private set; }
     public long ProductId { get; private set; }
 
     public ExternalPropertiesProductImageDTO() { }
 
-    public ExternalPropertiesProductImageDTO(string fileName, decimal fileLength, long productId)
+    public ExternalPropertiesProductImageDTO(string fileName, long fileLength, long productId)
     {
         FileName = fileName;
         FileLength = fileLength;
